Move handlers into separate package
